Make it clearer when product or test code uses libgit2 vs. managed
