+++ src/NerdBank.GitVersioning.Tests/RepoTestBase.Helpers.cs
@@ -74,7 +74,7 @@ public partial class RepoTestBase
     /// </remarks>
     protected System.Version GetIdAsVersion(string repoRelativeProjectDirectory = null)
     {
-        VersionOracle oracle = new LibGit2VersionOracle(
+        VersionOracle oracle = new Nerdbank.GitVersioning.LibGit2.LibGit2VersionOracle(
             repoRelativeProjectDirectory == null ? this.Repo.Info.WorkingDirectory : Path.Combine(this.Repo.Info.WorkingDirectory, repoRelativeProjectDirectory),
             this.Repo,
             null);
@@ -89,7 +89,7 @@ public partial class RepoTestBase
 
     protected static System.Version GetIdAsVersion(Repository repository, Commit commit, string repoRelativeProjectDirectory = null)
     {
-        VersionOracle oracle = new LibGit2VersionOracle(
+        VersionOracle oracle = new Nerdbank.GitVersioning.LibGit2.LibGit2VersionOracle(
             repoRelativeProjectDirectory == null ? repository.Info.WorkingDirectory : Path.Combine(repository.Info.WorkingDirectory, repoRelativeProjectDirectory),
             repository,
             commit,
